Edit Timer Digits

+++ lib/presenter/home/home_presenter_componant.dart
@@ -5,5 +5,5 @@ class HomePresenterComponant{
 
   void setView(HomeViewComponant componant){}
 
-  void moveToNextScreen(BuildContext context , int price , int timer , int speed){}
+  void moveToNextScreen(BuildContext context , double price , double timer , int speed){}
 }
